Modernized go fmt, lint etc fixes. Simple code cleanup
